DEP {DEP}	R Documentation

DEP: A package for Differential Enrichment analysis of Proteomics data.

Description

This package provides an integrated analysis workflow for robust and reproducible analysis of mass spectrometry proteomics data for differential protein expression or differential enrichment. It requires tabular input (e.g. txt files) as generated by quantitative analysis softwares of raw mass spectrometry data, such as MaxQuant or IsobarQuant. Functions are provided for data preparation, filtering, variance normalization and imputation of missing values, as well as statistical testing of differentially enriched / expressed proteins. It also includes tools to check intermediate steps in the workflow, such as normalization and missing values imputation. Finally, visualization tools are provided to explore the results, including heatmap, volcano plot and barplot representations. For scientists with limited experience in R, the package also entails wrapper functions that entail the complete analysis workflow and generate a report. Even easier to use are the interactive Shiny apps that are provided by the package.

Shiny apps

run_app: Shiny apps for interactive analysis.

Workflow functions

LFQ: Label-free quantification (LFQ) workflow wrapper.

TMT: Tandem-mass-tags (TMT) workflow wrapper.

report: Create a rmarkdown report wrapper.

Wrapper functions

import_MaxQuant: Import data from MaxQuant into a SummarizedExperiment object.

import_IsobarQuant: Import data from IsobarQuant into a SummarizedExperiment object.

process: Perform filtering, normalization and imputation on protein data.

analyze_dep: Differential protein expression analysis.

plot_all: Visualize the results in different types of plots.

Main functions

make_unique: Generate unique names.

make_se_parse: Turn data.frame into SummarizedExperiment by parsing column names.

make_se: Turn data.frame into SummarizedExperiment using an experimental design.

filter_proteins: Filter proteins based on missing values.

normalize_vsn: Normalize data using vsn.

impute: Impute missing values.

test_diff: Differential enrichment analysis.

add_rejections: Mark significant proteins.

get_results: Generate a results table.

Visualization functions

plot_single: Barplot for a protein of interest.

plot_volcano: Volcano plot for a specified contrast.

plot_heatmap: Heatmap of all significant proteins.

plot_normalization: Boxplots to inspect normalization.

plot_detect: Density and CumSum plots of proteins with and without missing values.

plot_imputation: Density plots to inspect imputation.

plot_missval: Heatmap to inspect missing values.

plot_numbers: Barplot of proteins identified.

plot_frequency: Barplot of protein identification overlap between conditions.

plot_coverage: Barplot of the protein coverage in conditions.

plot_pca: PCA plot of top variable proteins.

plot_cor: Plot correlation matrix.

plot_cor: Plot Gower's distance matrix.

plot_p_hist: P value histogram.

plot_cond_freq: Barplot of the number of significant conditions per protein.

plot_cond_overlap: Barplot of the number of proteins for overlapping conditions.

plot_cond: Barplot of the frequency of significant conditions per protein and the overlap in proteins between conditions.

Gene Set Enrichment Analysis functions

test_gsea: Gene Set Enrichment Analysis using enrichR.

plot_gsea: Barplot of enriched gene sets.

Additional functions

get_df_wide: Generate a wide data.frame from a SummarizedExperiment.

get_df_long: Generate a long data.frame from a SummarizedExperiment.

se2msn: SummarizedExperiment object to MSnSet object conversion.

filter_missval: Filter on missing values.

manual_impute: Imputation by random draws from a manually defined distribution.

get_prefix: Obtain the longest common prefix.

get_suffix: Obtain the longest common suffix.

Example data

UbiLength: Ubiquitin interactors of different linear ubiquitin lengths (UbIA-MS dataset) (Zhang, Smits, van Tilburg et al. Mol. Cell 2017).

UbiLength_ExpDesign: Experimental design of the UbiLength dataset.

DiUbi: Ubiquitin interactors for different diubiquitin-linkages (UbIA-MS dataset) (Zhang, Smits, van Tilburg et al. Mol. Cell 2017).

DiUbi_ExpDesign: Experimental design of the DiUbi dataset.

[Package DEP version 1.16.0 Index]