DiUbi {DEP}	R Documentation

DiUbi - Ubiquitin interactors for different diubiquitin-linkages (UbIA-MS dataset)

Description

The DiUbi dataset contains label free quantification (LFQ) and intensity-based absolute quantification (iBAQ) data for ubiquitin interactors of different diubiquitin-linkages, generated by Zhang et al 2017. The dataset contains the proteingroups output file from MaxQuant.

Usage

DiUbi

Format

A data.frame with 4071 observations and 102 variables:

Protein.IDs

Uniprot IDs

Majority.protein.IDs

Uniprot IDs of major protein(s) in the protein group

Protein.names

Full protein names

Gene.names

Gene name

Fasta.headers

Header as present in the Uniprot fasta file

Peptides

Number of peptides identified for this protein group

Razor...unique.peptides

Number of peptides used for the quantification of this protein group

Unique.peptides

Number of peptides identified which are unique for this protein group

Intensity columns (30)

Raw mass spectrometry intensity, A.U.

iBAQ columns (30)

iBAQ normalized mass spectrometry intensity, A.U.

LFQ.intensity columns (30)

LFQ normalized mass spectrometry intensity, A.U.

Only.identified.by.site

The protein is only identified by a modification site if marked ('+')

Reverse

The protein is identified in the decoy database if marked ('+')

Potential.contaminant

The protein is a known contaminant if marked ('+')

id

The protein group ID

Value

A data.frame.

Source

Zhang, Smits, van Tilburg, et al (2017). An interaction landscape of ubiquitin signaling. Molecular Cell 65(5): 941-955. doi: 10.1016/j.molcel.2017.01.004.

[Package DEP version 1.16.0 Index]