TMT {DEP}	R Documentation

TMT workflow

Description

TMT is a wrapper function running the entire differential enrichment/expression analysis workflow for TMT-based proteomics data. The protein table from IsobarQuant is used as direct input.

Usage

TMT(proteins, expdesign, fun = c("man", "bpca", "knn", "QRILC", "MLE",
  "MinDet", "MinProb", "min", "zero", "mixed", "nbavg"), type = c("all",
  "control", "manual"), control = NULL, test = NULL,
  name = "gene_name", ids = "protein_id", alpha = 0.05, lfc = 1)

Arguments

proteins
Data.frame, The data object.

expdesign
Data.frame, The experimental design object.

fun
"man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero", "mixed" or "nbavg", Function used for data imputation based on manual_impute and impute.

type
'all', 'control' or 'manual', The type of contrasts that will be generated.

control
Character(1), The sample name to which the contrasts are generated (the control sample would be most appropriate).

test
Character, The contrasts that will be tested if type = "manual". These should be formatted as "SampleA_vs_SampleB" or c("SampleA_vs_SampleC", "SampleB_vs_SampleC").

name
Character(1), Name of the column representing gene names.

ids
'Character(1), Name of the column representing protein IDs.

alpha
Numeric(1), sets the false discovery rate threshold.

lfc
Numeric(1), sets the log fold change threshold.

Value

A list of 8 objects:

se
SummarizedExperiment object containing the original data

filt
SummarizedExperiment object containing the filtered data

norm
SummarizedExperiment object containing the normalized data

imputed
SummarizedExperiment object containing the imputed data

diff
SummarizedExperiment object containing FDR estimates of differential expression

dep
SummarizedExperiment object annotated with logical columns indicating significant proteins

results
data.frame containing containing all results variables from the performed analysis

param
data.frame containing the test parameters

Examples

## Not run: 

TMT_res <- TMT()

## End(Not run)

[Package DEP version 1.16.0 Index]