add_rejections {DEP}	R Documentation

Mark significant proteins

Description

add_rejections marks significant proteins based on defined cutoffs.

Usage

add_rejections(diff, alpha = 0.05, lfc = 1)

Arguments

diff
SummarizedExperiment, Proteomics dataset on which differential enrichment analysis has been performed (output from test_diff()).

alpha
Numeric(1), Sets the threshold for the adjusted P value.

lfc
Numeric(1), Sets the threshold for the log2 fold change.

Value

A SummarizedExperiment object annotated with logical columns indicating significant proteins.

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Test for differentially expressed proteins
diff <- test_diff(imputed, "control", "Ctrl")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

[Package DEP version 1.16.0 Index]