analyze_dep {DEP}	R Documentation

Differential expression analysis

Description

analyze_dep tests for differential expression of proteins based on protein-wise linear models and empirical Bayes statistics using limma.

Usage

analyze_dep(se, type = c("all", "control", "manual"), control = NULL,
  alpha = 0.05, lfc = 1, test = NULL, design_formula = formula(~0 +
  condition))

Arguments

se
SummarizedExperiment, Proteomics data with unique names and identifiers annotated in 'name' and 'ID' columns. Additionally, the colData should contain sample annotation including 'label', 'condition' and 'replicate' columns. The appropriate columns and objects can be generated using make_se or make_se_parse.

type
"all", "control" or "manual", The type of contrasts that will be tested. This can be all possible pairwise comparisons ("all"), limited to the comparisons versus the control ("control"), or manually defined contrasts ("manual").

control
Character(1), The condition to which contrasts are generated (a control condition would be most appropriate).

alpha
Numeric(1), Sets the threshold for the adjusted P value.

lfc
Numeric(1), Sets the threshold for the log2 fold change.

test
Character, The contrasts that will be tested if type = "manual". These should be formatted as "SampleA_vs_SampleB" or c("SampleA_vs_SampleC", "SampleB_vs_SampleC").

design_formula
Formula, Used to create the design matrix.

Value

A SummarizedExperiment object containing FDR estimates of differential expression and logical columns indicating significant proteins.

Examples

# Load datasets
data <- UbiLength
exp_design <- UbiLength_ExpDesign

# Import and process data
se <- import_MaxQuant(data, exp_design)
processed <- process(se)

# Differential protein expression analysis
dep <- analyze_dep(processed, "control", "Ctrl")
dep <- analyze_dep(processed, "control", "Ctrl",
    alpha = 0.01, lfc = log2(1.5))
dep <- analyze_dep(processed, "manual", test = c("Ubi6_vs_Ubi4"))

[Package DEP version 1.16.0 Index]