filter_missval {DEP}	R Documentation

Filter on missing values

Description

filter_missval filters a proteomics dataset based on missing values. The dataset is filtered for proteins that have a maximum of 'thr' missing values in at least one condition.

Usage

filter_missval(se, thr = 0)

Arguments

se
SummarizedExperiment, Proteomics data (output from make_se() or make_se_parse()).

thr
Integer(1), Sets the threshold for the allowed number of missing values in at least one condition.

Value

A filtered SummarizedExperiment object.

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter
stringent_filter <- filter_missval(se, thr = 0)
less_stringent_filter <- filter_missval(se, thr = 1)

[Package DEP version 1.16.0 Index]