import_IsobarQuant {DEP}	R Documentation

Import from IsobarQuant

Description

import_IsobarQuant imports a protein table from IsobarQuant and converts it into a SummarizedExperiment object.

Usage

import_IsobarQuant(proteins, expdesign, intensities = "signal_sum",
  names = "gene_name", ids = "protein_id", delim = "[|]")

Arguments

proteins
Data.frame, Protein table for which unique names will be created.

expdesign
Data.frame, Experimental design with 'label', 'condition' and 'replicate' information. See UbiLength_ExpDesign for an example experimental design.

intensities
Character(1), Prefix of the columns containing sample intensities.

names
Character(1), Name of the column containing feature names.

ids
Character(1), Name of the column containing feature IDs.

delim
Character(1), Sets the delimiter separating the feature names within on protein group.

Value

A SummarizedExperiment object with log2-transformed values and "name" and "ID" columns containing unique names and identifiers.

Examples

## Not run: 
# Load data
isobarquant_table <- read.csv("testfile.txt", header = TRUE,
                              stringsAsFactors = FALSE, sep = "\t")
exp_design <- read.csv("test_experimental_design.txt", header = TRUE,
                              stringsAsFactors = FALSE, sep = "\t")
# Import data
se <- import_IsobarQuant(isabarquant_table, exp_design)

## End(Not run)

[Package DEP version 1.16.0 Index]