make_se {DEP}	R Documentation

Data.frame to SummarizedExperiment object conversion using an experimental design

Description

make_se creates a SummarizedExperiment object based on two data.frames: the protein table and experimental design.

Usage

make_se(proteins_unique, columns, expdesign)

Arguments

proteins_unique
Data.frame, Protein table with unique names annotated in the 'name' column (output from make_unique()).

columns
Integer vector, Column numbers indicating the columns containing the assay data.

expdesign
Data.frame, Experimental design with 'label', 'condition' and 'replicate' information. See UbiLength_ExpDesign for an example experimental design.

Value

A SummarizedExperiment object with log2-transformed values.

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

[Package DEP version 1.16.0 Index]