make_unique {DEP}	R Documentation

Make unique names

Description

make_unique generates unique identifiers for a proteomics dataset based on "name" and "id" columns.

Usage

make_unique(proteins, names, ids, delim = ";")

Arguments

proteins
Data.frame, Protein table for which unique names will be created.

names
Character(1), Name of the column containing feature names.

ids
Character(1), Name of the column containing feature IDs.

delim
Character(1), Sets the delimiter separating the feature names within one protein group.

Value

A data.frame with the additional variables "name" and "ID" containing unique names and identifiers, respectively.

Examples

# Load example
data <- UbiLength

# Check colnames and pick the appropriate columns
colnames(data)
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

[Package DEP version 1.16.0 Index]