meanSdPlot {DEP}	R Documentation

Plot row standard deviations versus row means

Description

meanSdPlot generates a hexagonal heatmap of the row standard deviations versus row means from SummarizedExperiment objects. See meanSdPlot.

Usage

meanSdPlot(x, ranks = TRUE, xlab = ifelse(ranks, "rank(mean)", "mean"),
  ylab = "sd", pch, plot = TRUE, bins = 50, ...)

Arguments

x
SummarizedExperiment, Data object.

ranks
Logical, Whether or not to plot the row means on the rank scale.

xlab
Character, x-axis label.

ylab
Character, y-axis label.

pch
Ignored - exists for backward compatibility.

plot
Logical, Whether or not to produce the plot.

bins
Numeric vector, Data object before normalization.

...
Other arguments, Passed to stat_binhex.

Value

A scatter plot of row standard deviations versus row means(generated by stat_binhex)

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter and normalize
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)

# Plot meanSdPlot
meanSdPlot(norm)

[Package DEP version 1.16.0 Index]