plot_all {DEP}	R Documentation

Visualize the results in different types of plots

Description

plot_all visualizes the results of the differential protein expression analysis in different types of plots. These are (1) volcano plots, (2) heatmaps, (3) single protein plots, (4) frequency plots and/or (5) comparison plots.

Usage

plot_all(dep, plots = c("volcano", "heatmap", "single", "freq",
  "comparison"))

Arguments

dep
SummarizedExperiment, Data object which has been generated by analyze_dep or the combination of test_diff and add_rejections.

plots
"volcano", "heatmap", "single", "freq" and/or "comparison",

Value

Pdfs containg the desired plots.

Examples

# Load datasets
data <- UbiLength
exp_design <- UbiLength_ExpDesign

# Import and process data
se <- import_MaxQuant(data, exp_design)
processed <- process(se)

# Differential protein expression analysis
dep <- analyze_dep(processed, "control", "Ctrl")

## Not run: 
# Plot all plots
plot_all(dep)

## End(Not run)

[Package DEP version 1.16.0 Index]