plot_cor {DEP}	R Documentation

Plot correlation matrix

Description

plot_cor generates a Pearson correlation matrix.

Usage

plot_cor(dep, significant = TRUE, lower = -1, upper = 1,
  pal = "PRGn", pal_rev = FALSE, indicate = NULL, font_size = 12,
  plot = TRUE, ...)

Arguments

dep
SummarizedExperiment, Data object for which differentially enriched proteins are annotated (output from test_diff() and add_rejections()).

significant
Logical(1), Whether or not to filter for significant proteins.

lower
Integer(1), Sets the lower limit of the color scale.

upper
Integer(1), Sets the upper limit of the color scale.

pal
Character(1), Sets the color panel (from RColorBrewer).

pal_rev
Logical(1), Whether or not to invert the color palette.

indicate
Character, Sets additional annotation on the top of the heatmap based on columns from the experimental design (colData).

font_size
Integer(1), Sets the size of the labels.

plot
Logical(1), If TRUE (default) the correlation matrix plot is produced. Otherwise (if FALSE), the data which the correlation matrix plot is based on are returned.

...
Additional arguments for Heatmap function as depicted in Heatmap

Value

A heatmap plot (generated by Heatmap)

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Test for differentially expressed proteins
diff <- test_diff(imputed, "control", "Ctrl")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

# Plot correlation matrix
plot_cor(dep)

[Package DEP version 1.16.0 Index]