plot_gsea {DEP}	R Documentation

Plot enriched Gene Sets

Description

plot_gsea plots enriched gene sets from Gene Set Enrichment Analysis.

Usage

plot_gsea(gsea_results, number = 10, alpha = 0.05, contrasts = NULL,
  databases = NULL, nrow = 1, term_size = 8)

Arguments

gsea_results
Data.frame, Gene Set Enrichment Analysis results object. (output from test_gsea()).

number
Numeric(1), Sets the number of enriched terms per contrast to be plotted.

alpha
Numeric(1), Sets the threshold for the adjusted P value.

contrasts
Character, Specifies the contrast(s) to plot. If 'NULL' all contrasts will be plotted.

databases
Character, Specifies the database(s) to plot. If 'NULL' all databases will be plotted.

nrow
Numeric(1), Sets the number of rows for the plot.

term_size
Numeric(1), Sets the text size of the terms.

Value

A barplot of the enriched terms (generated by ggplot).

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Test for differentially expressed proteins
diff <- diff <- test_diff(imputed, "control", "Ctrl")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

## Not run: 

# Test enrichments
gsea_results <- test_gsea(dep)
plot_gsea(gsea_results)

## End(Not run)

[Package DEP version 1.16.0 Index]