plot_missval {DEP}	R Documentation

Plot a heatmap of proteins with missing values

Description

plot_missval generates a heatmap of proteins with missing values to discover whether values are missing by random or not.

Usage

plot_missval(se)

Arguments

se
SummarizedExperiment, Data object with missing values.

Value

A heatmap indicating whether values are missing (0) or not (1) (generated by Heatmap).

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)

# Plot missing values heatmap
plot_missval(filt)

[Package DEP version 1.16.0 Index]