plot_normalization {DEP}	R Documentation

Visualize normalization

Description

plot_normalization generates boxplots of all conditions for input objects, e.g. before and after normalization.

Usage

plot_normalization(se, ...)

Arguments

se
SummarizedExperiment, Data object, e.g. before normalization (output from make_se() or make_se_parse()).

...
Additional SummarizedExperiment object(s), E.g. data object after normalization (output from normalize_vsn).

Value

Boxplots of all conditions for input objects, e.g. before and after normalization (generated by ggplot). Adding components and other plot adjustments can be easily done using the ggplot2 syntax (i.e. using '+')

Examples

# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter and normalize
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)

# Plot normalization
plot_normalization(se, filt, norm)

[Package DEP version 1.16.0 Index]