process {DEP}	R Documentation

Proteomics data processing

Description

process performs data processing on a SummarizedExperiment object. It (1) filters a proteomics dataset based on missing values, (2) applies variance stabilizing normalization and (3) imputes eventual remaining missing values.

Usage

process(se, thr = 0, fun = c("man", "bpca", "knn", "QRILC", "MLE",
  "MinDet", "MinProb", "min", "zero", "mixed", "nbavg"), ...)

Arguments

se
SummarizedExperiment, Proteomics data with unique names and identifiers annotated in 'name' and 'ID' columns. The appropriate columns and objects can be generated using the wrapper import functions import_MaxQuant and import_IsobarQuant or the generic functions make_se and make_se_parse.

thr
Integer(1), Sets the threshold for the allowed number of missing values per condition.

fun
"man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero", "mixed" or "nbavg", Function used for data imputation based on manual_impute and impute.

...
Additional arguments for imputation functions as depicted in manual_impute and impute.

Value

A filtered, normalized and imputed SummarizedExperiment object.

Examples

# Load datasets
data <- UbiLength
exp_design <- UbiLength_ExpDesign

# Import data
se <- import_MaxQuant(data, exp_design)

# Process data
processed <- process(se)

[Package DEP version 1.16.0 Index]