report {DEP}	R Documentation

Generate a markdown report

Description

report generates a report of the analysis performed by TMT and LFQ wrapper functions. Additionally, the results table is saved as a tab-delimited file.

Usage

report(results)

Arguments

results
List of SummarizedExperiment objects obtained from the LFQ or TMT wrapper functions.

Value

A rmarkdown report is generated and saved. Additionally, the results table is saved as a tab-delimited txt file.

Examples

## Not run: 

data <- UbiLength
expdesign <- UbiLength_ExpDesign

results <- LFQ(data, expdesign, 'MinProb', 'control', 'Ctrl')
report(results)

## End(Not run)

[Package DEP version 1.16.0 Index]